Atualiza teste de criação de usuário
